remove variant
