--- packages/office-ui-fabric-react/src/components/Callout/AnnouncedCallout.types.ts
@@ -1,3 +1,0 @@
-import { ICalloutProps } from './Callout.types';
-
-export interface IAnnouncedCalloutProps extends ICalloutProps {}
