+++ packages/office-ui-fabric-react/src/components/Callout/Callout.doc.tsx
@@ -3,13 +3,13 @@ import * as React from 'react';
 import { IDocPageProps } from '../../common/DocPage.types';
 
 import { CalloutBasicExample } from './examples/Callout.Basic.Example';
-import { AnnouncedCalloutExample } from './examples/Callout.Announced.Example';
+import { StatusCalloutExample } from './examples/Callout.Status.Example';
 import { CalloutFocusTrapExample } from './examples/Callout.FocusTrap.Example';
 import { CalloutDirectionalExample } from './examples/Callout.Directional.Example';
 import { CalloutCoverExample } from './examples/Callout.Cover.Example';
 
 const CalloutBasicExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.Basic.Example.tsx') as string;
-const AnnouncedCalloutExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.Announced.Example.tsx') as string;
+const StatusCalloutExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.Status.Example.tsx') as string;
 const CalloutFocusTrapExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.FocusTrap.Example.tsx') as string;
 const CalloutDirectionalExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.Directional.Example.tsx') as string;
 const CalloutCoverExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Callout/examples/Callout.Cover.Example.tsx') as string;
@@ -31,9 +31,9 @@ export const CalloutPageProps: IDocPageProps = {
       view: <CalloutFocusTrapExample />
     },
     {
-      title: 'AnnouncedCallout Variant',
-      code: AnnouncedCalloutExampleCode,
-      view: <AnnouncedCalloutExample />
+      title: 'Non Focused Callout Example',
+      code: StatusCalloutExampleCode,
+      view: <StatusCalloutExample />
     },
     {
       title: 'Callout with directional hint',
