+++ packages/office-ui-fabric-react/src/components/Callout/docs/CalloutOverview.md
@@ -1,6 +1,6 @@
 Callouts are a powerful way to simplify a user interface. They host tips and other information users need when they need it, with minimal effort on their part. Callouts can help you use screen space more effectively and reduce screen clutter. However, poorly designed Callouts can be annoying, distracting, unhelpful, overwhelming, or in the way.
 
-Callouts that do not contain focusable elements (links, buttons etc) cannot gain focus using `setInitialFocus`. For proper screen reader support use the `AnnouncedCallout` variant which treats the callout content like a status message.
+Callouts that do not contain focusable elements (links, buttons etc) cannot gain focus using `setInitialFocus`. For proper screen reader support follow the `Non Focused Callout` example which treats the callout content like a status message.
 
 Use a Callout for displaying additional contextual information about an item on the screen. Callouts also have a tail that identifies their source. A common use for Callout is the introduction of a new feature or capability of an app or site. Alternate usages include pairing the Callout with a button or clickable element for on-demand presentation of additional or supporting content.
