+++ packages/office-ui-fabric-react/src/components/Callout/index.ts
@@ -2,6 +2,4 @@ export * from './Callout';
 export * from './Callout.types';
 export * from './FocusTrapCallout';
 export * from './FocusTrapCallout.types';
-export * from './AnnouncedCallout';
-export * from './AnnouncedCallout.types';
 export * from '../../common/DirectionalHint';
